Change returns to Option<T>
Changed the returns to option<T> because I couldnt retrn self as it
was just &mut. Need a way to call adv_cycles while taking ownership
of self or need to use mem::take() (I think) and implement default.
Both of which are more annoying right now so option<T> it is