Modify how dma reads and writes
Its possible that dma could use vram as a source we need to be able
to read from both ppu and memory as a source. When doing dma, the
state of the ppu should not interfere with whether the read or write
is successful so created new interface to do the reads and writes.
When inside the ppu states, they will still need special logic to
decide if the read/write should go through or not