Update the four shades of grey
want to color the screen in shades of red for fun so changed
the rgba values to be different reds. Also realized that sdl is
reading the values as BGRA8888 rather than ARGB8888 like specified
in the pixelformatenum. Guessing this is because im on little endian
it also expects the input as little endian