Fixed scrolling problems on pokemon intro
scroll registers should be frozen at the beggining of the fifo
fetch phase. The lcdc register should also be frozen at the begining
of the fifo fetch phase.
Also needed to store the lower 3 bits of the scx register for the
entirety of a scanline as they should be constant for a scanline